fix: reorganized components
