+++ src/components/LobbyView/LobbyPlayerIndicator.vue
