+++ src/components/RulesView/RulePreview.vue
