+++ src/components/StatsView/AwardCard.vue
