+++ src/components/StatsView/StatsScoringDialog.vue
